refactor(kff.createObject, kff.View): get rid of __proto__ and Object.setPrototypeOf
Direct setting of __proto__ is controversial and it doesn't make significant performance boots, it is even slower than copying properties from one object to another in some cases. Object.setPrototypeOf is faster in Chrome, but slow in FF (and it's only present in fresh nightly builds of Chrome and FF). Chrome is fast anyway, so I prefer getting better performance in FF with "safe" programming methods.