refactor(kff.View, kff.BindingView): remove setParentView calls from clone methods
setParentView method was called multiple times in deep cloned view hierarchies because each cloning involved cloning in nested views as well and then again when cloning subviews etc. The solution is to remove any redundant calls in clone method (even reseting parentView in constructor to null) and call it once only on root of cloned view tree in BindingView.