refactor(all): remove remaining dependencies on jQuery and add simple $ wrapper as a substitute
Now it is possible to run framework without jQuery. By default, however, if window.jQuery exists, it will be used, if it does not, simple wrapper will be used instead. It is also possible to force use of simple wrapper instead of jQuery even if it exists by setting kff.useJQuery = false after kff is loaded and before any View is rendered. For now, this is just experimental support. IE8 is supported with provided ie8polyfills.js (standard DOM listeners and classList polyfill).