feat(kff.BindingView): implement live sorting into collection binding
Live sorting (sorting of bound collection) is done in similar way as filtering. Sort function is specified in attribut 'data-kff-sort' using 'model.compareFunction' format. It watches any change in collection or in collection item, which can lead to an overhead in some situations but DOM manipulation is done only on actual change in order, so it should not be critical.