feat(kff.PageView): implement HTML regions
A new option called 'regions' is now available in page view. It should be a plain object that contains key-value pairs in form of 'css selector': 'html template string'. When the page view is rendered (before render method), each html template is inserted into the DOM node(s) matching the css selector. If the node is not empty, its content is cached in a temporary document fragment and it is inserted back to the node when the view is being destroyed. So the overlaying views can even overwrite regions of previous views without damaging entire app.