feat(kff): eliminate dependency for document and window objects in views
Removing dependencies for global browser objects is an important step for server side rendering in node.js environment. Document and window objects are now passed in context object as a option in views, app and front controller. Note that they are still used in polyfills, but these are not used in node.