feat(kff.Binder): implement 'dispatch' binder operator that passes DOM binding events to some sort of global dispatcher instead of pushing to the model directly (aka Flux pattern)
How it works: when the 'dispatch' operator is present in a binding, the value from the DOM change is not set on the model via two way binding, but is wrapped into event object that is passed to the dispatchEvent method of the view. The event will "bubble" up the view hierarchy and will try to call a dispatchEvent method on any view that implements it. The bubbling will stop whenever the dispatchEvent method on some view returns false.