refactor(kff.Dispatcher): dispatcher should handle sync and async actions as well
To determine if the action is sync, the dispatcher checks the function arity (length propetry). It should be either 1 or 2. Functions with arity 1 are considered sync and the dispatcher expect the next event to be directly returned. In case of arity 2, the function is considered async - the first argument is the event and the second argument is node-style error-first callback (err, event).