feat(View): a new template method shouldRefresh added
If a view contains a shouldRefresh method, it will be called before any refresh cycle. If it returns false, the refresh won't proceed on this view and all its subviews. It is useful for speed optimization of large views (for example list with thousands items) that spend too much time even in data comparison and evaluation.