refactor(ServiceContainer): change semantic of the "calls" declaration
BREAKING CHANGE: The calls declaration is now an array or array of arrays. The first item of array is the name of the method to be called and subsequent array items are the arguments of that method. This is much less verbose than the previous scheme (which was not really used in the wild)