refactor(InsertBinder): remove second (negate) and third (force) parameter of the if/ifnot binder
Negate parameter was redundant as it is embedded in the ifnot binder. The force parameter is the default behavior in both the if and ifnot binder and it is not desirable to change it (it is recommended to hide the element using the class binder instead to letting  the DOM element live outside of the document).