feat(CollectionBinder): support for keyed collection items
By default, equality of collection items is checked by comparing references. But in case of immutable data, references change whenever any property of the item changes. This problem can be avoided by using a new collection binding operator :key(keyProp), where keyProp is the name of key property. It uses keys for comparison instead of object references. Keys must be unique strings. Note that using keys is optional but it is strongly recommended when using animations.