refactor(all): remove ServiceContainer, Service, App classes and related code
BREAKING CHANGE: As the view constructors can be assigned directly to the scope of the main (top-level) view, there is currently no need for DI container. Also App class was only a leftover wrapper for the main view and so it was pointless. The app initialization code can now be reduced only to creating the main view instance and calling initAll method on it. This will break any application that depends on App or ServiceContainer, but that can be fixed easily.